+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### ✨ Enhancements
 
 - Updated `SnapshotTracer` to open the generated HTML report in a one-shot server, instead of opening the HTML file directly.
+- Extended infinite-loop check to flag while loops whose conditions depend only on immutable variables and are not reassigned in the loop body.
 - Added custom renderers for the PEP8 error codes: E204, E225, E231
 - Added custom renderers for the PEP8 error codes: E271, E274, E502